Extract voting publications actions to its own class
