+++ decidim-elections/app/controllers/decidim/votings/admin/votings_controller.rb
@@ -58,28 +58,6 @@ module Decidim
           end
         end
 
-        def publish
-          enforce_permission_to :publish, :voting, voting: current_voting
-
-          PublishVoting.call(current_voting, current_user) do
-            on(:ok) do
-              flash[:notice] = I18n.t("votings.publish.success", scope: "decidim.votings.admin")
-              redirect_to edit_voting_path(voting)
-            end
-          end
-        end
-
-        def unpublish
-          enforce_permission_to :unpublish, :voting, voting: current_voting
-
-          UnpublishVoting.call(current_voting, current_user) do
-            on(:ok) do
-              flash[:notice] = I18n.t("votings.unpublish.success", scope: "decidim.votings.admin")
-              redirect_to edit_voting_path(voting)
-            end
-          end
-        end
-
         def available_polling_officers
           respond_to do |format|
             format.json do
